cleanup abstraction between get and update auctions (code re-use)
